Refactoring: move ClonedErrata to parent package
