+++ java/code/src/com/redhat/rhn/frontend/action/errata/ChannelSetupAction.java
@@ -19,7 +19,7 @@ import static com.redhat.rhn.frontend.struts.RequestContext.ERRATA_ID;
 import com.redhat.rhn.common.db.datasource.DataResult;
 import com.redhat.rhn.domain.channel.Channel;
 import com.redhat.rhn.domain.errata.Errata;
-import com.redhat.rhn.domain.errata.impl.ClonedErrata;
+import com.redhat.rhn.domain.errata.ClonedErrata;
 import com.redhat.rhn.domain.rhnset.RhnSet;
 import com.redhat.rhn.domain.rhnset.RhnSetElement;
 import com.redhat.rhn.domain.user.User;
